make store finish its work provided that the session is active when it starts.
earlier, the store would stop working if the imap session closed while
store was working. which shouldn't happen but sometimes clients are
naughty. now, Store picks up a session when execute() is first called, and
uses that until it's done.

it would also be possible to have Store abort cleanly. I think I prefer to
finish.