save one SQL select per IMAP select (hi squirrelmail, hi roundcube).
as a side effect, this starts sending FLAGS and PERMANENTFLAGS updates,
which will mostly please imaptest. There's a potential problem: If an IMAP
STORE races againt an IMAP FETCH and the FETCH is blocked due to lock
contention, then the FETCH might complete before the flag update is sent.
in that case, the flag update migth be sent after the first fetch response
that uses the new flag.