convert thread-index to references.
this works fairly well, but doesn't work (ie. does nothing) in some cases
of incomplete threads. one hopes that such incomplete threads will not be
common.

this often sends a zero-row COPY to the database. I could write code to
not do that, but maybe it's better to write that code in Postgres, Query
or Transaction. Maybe a new function, bool Query::isNoop() const, and
Q::execute() and T::enqueue() both do if ( isNoop() ) return?