check that each Session is connected to a live Connection.
this should "fix" the horrible memory leak. no idea how, but somehow some
connections managed to close themselves and leave a Session still attached
to its Mailbox.

this change checks that a Session is sane before notifying it of new mail,
so when that bug hits, its effects will be cancelled next time the mailbox
is changed.