prepare for reference-counted GC.
the PT has a circular data structure, so it won't be cleaned up by a
refcounter unless someone helps. this change helps.

the pointers to Nodes are all in PT or PT::Iterator, so when the PT is
dropped (by ~PT or explicit clear()), then we can tear down all the Node
cycles. that may still leave Iterators pointing to Nodes, which is okay.
the Nodes still point to their data, so the Iterators remain valid.
however, when an iterator is ++'d or --'d, it will become null (and the
object it pointed to will be reclaimed).