Implement PgCancel support and fill in the Query::cancel() blank.
Query::cancel() calls Database::cancel( this ) if it is Executing and
canBeSlow. (The state can't be checked later, because Query::cancel()
sets it to Failed at the start.) Database::cancel() indiscriminately
calls ->cancel(q) on every handle. I could have introduced a virtual
isExecuting(q), but since Postgres has to run the d->queries.find(q)
either way, this was simpler.) Postgres::cancel(), if it finds the
relevant query in its queue, creates a PgCanceller object to issue
the cancel request on a new connection.

PgCanceller inherits from Postgres (so as to share the constructor), and
overrides react(Connect) to issue the cancel request and tear down the
connection. In principle, we should issue the request and wait for the
server to close the connection, but it doesn't seem worth the bother.

Not yet fully tested.