cache "ON [UNSEEN ..." to make SM and other webmail clients faster.
not perfect: Injecting a message will increase the modseq and cause a
cache miss. but I think that's okay. we do so much database work when the
insert the message that selecting for unseen won't hurt very much more.

or will it? we could avoid it in a common case. if modseq has increased by
one and uidnext by one since we wrote into firstunseencache, then we still
have a cache hit... maybe do that later.