make aox survive a postgres restart.
now, if pg reports a crash or administrative restart, aox shuts down its
connections (which may lead aox to shut down, separate issue) and when it
needs to do database work it will reconnect.

two issues:

if aox loses its last db connection and knows it can't reconnect, then it
will shut down at once.

if aox needs to reconnect and pg isn't back yet, it probably will die...
maybe it ought to have a bit of patience in that case. if connection is
refused, maybe it should wait a little and try again.