Decode "indefinite form" lengths in LDAP responses.
This is needed to interoperate with Active Directory.

The patch was contributed by Jason Weathered; applied with some minor
corrections (we make sure we have the entire length before we remove
anything, even the type byte, so that repeated calls to parse() with
incomplete input work correctly; and fail() on length errors).