change the SMTP idle timeout stuff
the previous version led to thrashing if the smarthost is some distance
away, delivery takes more than a second, and we have a queue. one
connection would be kept open, the others would be rapidly closed and
reopened.

this version keeps all connections open for 15 seconds after use, and one
open for as long as the SMTP specification allows.