allow rollback() until we know it's too late.
the previous commit refused to rollback() after we'd done commit().
however, that's wrong. we can rollback() until the server has finished
processing the commit.

this makes a difference for cases like this:

   t->enqueue( "insert into ... select", 0 );
   t->commit();

the query can fail, and if so, the commit does nothing, and in that case
we can sensibly rollback.