turn FETCH/SEARCH 1:9999999999999999 into 1:*
in the past, we'd send BAD for MSNs larger than *, as UW did and Mark
likes. now we turn big MSNs into *, which satisfies the android client and
is like dovecot.

we do send an untagged message noting the change.

maybe we should return UID(int_max) instead of *, so UID SEARCH would work
as the client appears to intend?