new feature: throttle the Fetcher to match output speed.
if an IMAP client issues FETCH 1:* RFC822 across a slow connection, the
Fetcher could retrieve the entire mailbox and buffer it in RAM. mailboxes
can be big.

after this change, the Fetcher can be told about an output Buffer, and
will refrain from issuing new database queries if the buffer is bigger
than a megabyte.

untested.