improve db timeout logic
the old logic was based on extendTimeout(), and would start with a
ten-second timeout and then add the db-handle-inteval to the timeout every
time it sent a query. that is, query number 5 would get 5*timeout+10
seconds to finish, minus whatever the first four needed (usually a second
or less in total).

this logic:
 - a query that canBeSlow() receives db-handle-interval, but at least 60
   seconds
 - other queries receive db-handle-interval, but at most 10 seconds
 - if a transaction is idle, not progressing, and we're using most/all
   handles, we roll that transaction back after 20 seconds idleness