fix transaction idleness timeout
the old code was called only if !usable(), and a transaction can be usable
even if its last query was in a transaction. this code is called anyway.

also, change the timeout so that a transaction can be Failed for only five
seconds, idle for up to 20. after 5 (or 20) seconds of complete inaction
it loses its database handle, boom.