use the cipher we prefer, not the one the client prefers
the client gives us a list of what it supports, listed in order of
preference. openssl too has a list of supported ciphers, in order of
preference. this change makes it pick the one its internal list says is
best.

axel rau's problem seems to be related to the cipher choices, so I just
try something.