do record a delivery if it happens.
it is possible that the database lock times out while we're transmitting
data to the smarthost. in that case, our transaction is gone, but the
messsage has been sent and we'd better record that.

so do it.

this is very, very evil. it would be much better to hold onto the lock for
long enough and/or grab the lock (again) just before sending the the
CRLF.CRLF