extend the SMTP server timeout if we're writing a large body.
sending large bodies to remote SMTP smarthosts can take a long time.  this
shouldn't happen, the smarthost should be close, but some fools do that
any way, e.g. the fool Arnt, whose smarthost is on the far end of an 1Mbps
connection.

therefore, this change checks, at timeout, whether we're writing something
and if so, whether we've made any progress. if we have, then we just log
some debug output and patiently wait for the transfer to the smarthost to
finish.