simpler and more reliable view updates
this should Just Work. but we'll see.

the old code attempted to optimise for fewer database transactions. but
it's better to just do them. since we initiate session updates from the db
anyway, this may be optimal, even without the hackery.

if we do need the hackery, then we should do it with new, separate
variables. having one variable store different data depending on whether
it's for a view or another mailbox is bad bad bad.