Fix sieve subaddress parsing when address-separator is empty.
According to the documentation, if address-separator is empty then the default behavior is to search for both '+' and '-' .  This was only implemented in one of two necessary places, so no match will occur when it would have been expected.

Since the code for the two occurances is almost the same, someone may want to refactor it.