backslash did need more treatment
this doesn't handle UTF-8 quite as 5738bis says; it wants to code to check
the UTF byte sequences first, then look for the whatsitcalled, the
terminating quote. I look for the quote instead and during that phase
treat any sequence of >=128 as likely UTF, then I check the UTF8 using
Codec later.