increase server-processes to 2 for better resistance against segfaults.
if a child dies, the mother will replace it with a new one. however, there
are sanity checks and blah, and if there's only one child the mother may a
little too easily decide that something is horribly wrong and quit.

when in reality something is horribly wrong: I think this is really due to
some crash fault in the server.