conditionally disable imap quota
Jean-Gérard is bothered by poor performance; some IMAP clients
ask about quotas too often and it's a lot of database work. it would
be possible to lessen the pain with a RAM cache, but if the quota
support is really a noop anyway, why bother? Noops don't have to be
fast, they can be removed.

this change leaves quota support enabled by default, since it's
nice to report if convenient and not too expensive. but it's now
easy to disable.