have a look at the state of the system after GC, and maybe die.
basically: each child of the master process looks at its memory status
regularly, and if it's using too much RAM or not being efficient, it lets
the master process replace itself, while being rather polite to the IMAP
clients who need to disconnect.

this is not complete, just a good commit step.