send fewer autoreplies by default
before, two messages inbound with different Subject would both receive an
autoreply, since the default handle depended on the subject. now, the
default handle is empty, so all messages from the same sender have the
same handle, and the number of messages to that recipient will be limited
more strictly by default.