Work around text/citext prepare limitation
Axel Rau had a message which 3.2.0 could not inject. I was able to
reproduce it, and together with Abhijit traced it down to a select which
succeeds on its own but not via prepare/execute.

For some reason, issuing "select * from addresses where domain='foo'"
is case-insensitive directly, but case-sensitive via prepare/execute.

This change forces insenstivity.