old message-id-less messages aren't in any db thread.
that shouldn't happen to new messages, but the old ones have to stay as
is. the thread_roots table is not friendly to id-less messages. so as of
this change, fetch x-gm-thrid reports 0 for such messages and no warning
is logged.

0 is a legal value according to google's spec (such as it is), and will
not conflict with any actual IDs. that all id-less messages are in the
same "thread" is... well... it could almost be seen as a feature.