log out idle clients faster.
broken NAT (even at ISPs) is growing increasingly common, and aox can have
problems with very large numbers of idle/gone connections from mobile
clients. the client gets a new IP address from the network and reconnects,
but when aox sends a few bytes to the old connection(s), the broken NAT
gateway doesn't tell aox that the connection is gone. and so the number of
connections increases, and increases, and increases. eventually aox runs
into resource problems.

this will clean up the idle connections after 30/60 minutes. imap clients
are required to be active at least every 29 minutes, so this is tolerably
conservative.