Don't check lockSanity at startup
We composed the query before the server_version was discovered, so the
test to decide whether to generate procpid or pid was always generating
procpid.

The point of the failing query was to warn users about locks at startup,
but the utility of the nice log message is outweighed by the price of
having to generate version-dependent queries.

(LockSpotter stays, because it runs much later.)