base 'younger' on delete time if searching deleted mail.
this takes care of to the presumably rather common use case "undelete
whatever has been deleted in the past day", which I just now assumed to
be what "aox undelete /users/arnt/inbox younger 1" does. it wasn't. noe
it is, as a matter of DWIM.

the previous behaviour used the original insertion time.

I've also updated the documentation on the web, but not pushed that
change.