Remove SSLv3 support.
As far as I can see, SMTP, IMAP and POP clients are essentially
invulnerable to the attack du jour, since the server (or an impostor)
can't really make clients open new connections. It would be a strange
client that lets a server tell it to open thousands connections and
control the preauth commands.

But I see no real reason to support SSLv3 any more anyway, so this change
disable SSLv3 support.