fix and reenable COMPRESS=DEFLATE
the problem was that zlib's allocated and then used it in ways our
allocator does not allow: mixing pointers and random binary data. if
that's done often enough, sooner or later the allocator will misinterpret
a random number as a pointer and then not free the object "pointed to".

this change leverages the existing fd/thread cleanup.