Adapt server_version parsing for Postgres 10+
The major version used to be 9.x, but with Postgres 10, the project has
switched to using x.y version numbers instead of x.y.z version numbers.
This means the major version is 10, and we can't expect the version
string to have a "section 2" at all.

We don't care about the minor version anyway, so we can decide what to
do based on the first section (9 or 10), at least until Postgres adopts
an Oracle-esque 12i versioning scheme.