SMTPUTF8 support at start time
This is a better default than the false, because it allows the first
outgoing message to use SMTPUTF8. If the relayhost doesn't, that'll bounce
(which is the same as would happen if the variable were false) and if the
upstream supports it all is well.

After the first connection to the relayhost, aox' outgoing SMTPUTF8
setting matches the relayhost's.