Generate BODY responses according to the database content
This is tricky: UTF8=ACCEPT and IMAP4rev2 change the way the bodystructure
is reported a little bit. message/global is treated like message/rfc822.
However, the database content depends on whether message/global was
treated like message/rfc822 at injection time.

This change makes the IMAP FETCH command report what's in the database. If
the database contents treat message/global like message/rfc822 and the
client can accept itq, then that's reported that way. If it's stored in
the old way or the client accepts only the old way, then it's reported as
RFC3501 requires.